requirements.txt
flask==2.1.3
requests~=2.28.1
Pillow~=9.2.0